# docs/api/new-api-asset-assessment.openapi.yaml
openapi: 3.0.3
info:
  title: new-api Asset Pre-evaluation and Assessment API
  version: 1.0.0
  description: |
    Browser-facing contract implemented by new-api. Downstream GPUFabric and
    asset-assessment-service internal APIs are documented separately.
servers:
  - url: /
...
paths:
...
  /api/banking/provider/pre-evaluations/{taskRef}:
    parameters:
      - $ref: '#/components/parameters/TaskRef'
    get:
      tags: [Banking Pre-evaluations]
      summary: Get a pre-evaluation task
      operationId: getPreEvaluation
      parameters:
        - $ref: '#/components/parameters/NewApiUser'
      responses:
        '200':
          description: Task projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreEvaluationEnvelope'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/banking/provider/pre-evaluations/{taskRef}/evidence:
    parameters:
      - $ref: '#/components/parameters/TaskRef'
    post:
      tags: [Banking Pre-evaluations]
      summary: Submit raw offline collector JSON
      description: |
        The body is limited to 4 MiB and is forwarded without reserialization.
        It must never be persisted or logged by new-api.
      operationId: submitOfflineCollectorEvidence
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/NewApiUser'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Evidence accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreEvaluationEnvelope'
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/banking/provider/pre-evaluations/{taskRef}/html:
    parameters:
      - $ref: '#/components/parameters/TaskRef'
    get:
      tags: [Banking Pre-evaluations]
      summary: Legacy path alias returning a verified PDF
      operationId: getPreEvaluationHtml
      parameters:
        - $ref: '#/components/parameters/NewApiUser'
      responses:
        '200':
          description: Rendered PDF report (legacy path alias)
          headers:
            Cache-Control:
              schema: {type: string, example: 'private, no-store'}
            X-Content-SHA256:
              schema: {$ref: '#/components/schemas/Sha256'}
          content:
            application/pdf:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/banking/provider/pre-evaluations/{taskRef}/pdf:
    parameters:
      - $ref: '#/components/parameters/TaskRef'
    get:
      tags: [Banking Pre-evaluations]
      summary: Get a verified PDF pre-evaluation report
      operationId: getPreEvaluationPdf
      parameters:
        - $ref: '#/components/parameters/NewApiUser'
      responses:
        '200':
          description: Rendered PDF report
          headers:
            Cache-Control:
              schema: {type: string, example: 'private, no-store'}
            X-Content-SHA256:
              schema: {$ref: '#/components/schemas/Sha256'}
          content:
            application/pdf:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
security:
  - UserBearer: []
  - UserSession: []
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
      description: Existing new-api user access token. New-Api-User is also required.
    UserSession:
      type: apiKey
      in: cookie
      name: session
      description: Existing new-api authenticated session cookie.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    NewApiUser:
      name: New-Api-User
      in: header
      required: true
      description: Required by existing new-api UserAuth middleware for both session and Bearer authentication.
      schema: {type: string}
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema: {type: string, maxLength: 128}
...
    TaskRef:
      name: taskRef
      in: path
      required: true
      schema: {type: string, minLength: 8, maxLength: 64}
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, code, message, data, requestId]
      properties:
        success: {type: boolean, enum: [true]}
        code: {type: string}
        message: {type: string}
        data: {}
        requestId: {type: string}
    ErrorEnvelope:
      type: object
      required: [success, code, message, retryable, requestId]
      properties:
        success: {type: boolean, enum: [false]}
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
        details: {type: object, additionalProperties: true}
        requestId: {type: string}
...
    PreEvaluationStatus:
      type: string
      enum: [requested, collecting, generating, generated, failed, expired]
    PreEvaluationReport:
      type: object
      required: [reportId, schemaVersion, reportSha256, hashProfile]
      properties:
        reportId: {type: string}
        schemaVersion: {type: string, enum: [gpuf.pre_evaluation.v1]}
        reportSha256: {$ref: '#/components/schemas/Sha256'}
        reportHtmlSha256: {$ref: '#/components/schemas/Sha256'}
        hashProfile: {type: string}
        technicalSnapshotId: {type: string}
        technicalSnapshotSha256: {$ref: '#/components/schemas/Sha256'}
        technicalSnapshotSchemaVersion: {type: string}
    PreEvaluation:
      type: object
      required: [taskRef, assetRef, sourceType, status, createdAt]
      properties:
        taskRef: {type: string}
        assetRef: {type: string}
        sourceType: {type: string, enum: [gpuf_online, offline_collector]}
        status: {$ref: '#/components/schemas/PreEvaluationStatus'}
        report:
          allOf: [{$ref: '#/components/schemas/PreEvaluationReport'}]
          nullable: true
        summary: {type: object, additionalProperties: true}
        error:
          allOf: [{$ref: '#/components/schemas/SafeError'}]
          nullable: true
        createdAt: {type: string, format: date-time}
        completedAt: {type: string, format: date-time, nullable: true}
    PreEvaluationEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: {$ref: '#/components/schemas/PreEvaluation'}
...
    SafeError:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
    Sha256:
      type: string
      pattern: '^[0-9a-f]{64}$'
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Unauthorized:
      description: Authentication or callback signature is missing
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    NotFound:
      description: Resource does not exist or is not owned by the caller
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Conflict:
      description: Idempotency, state, ownership binding, or event conflict
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    PayloadTooLarge:
      description: Request body is too large
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    BadGateway:
      description: Upstream response or integrity validation failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Unavailable:
      description: Required upstream or security configuration is unavailable
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...